Add new stats page for api usage #802
