+++ monitor/urls.py
@@ -32,12 +32,16 @@ urlpatterns = [
     url(r'^stats/$',
         login_required(TemplateView.as_view(template_name='monitor/stats.html')),
         name='monitor-stats'),
+    url(r'^stats/api$', TemplateView.as_view(template_name='monitor/api_usage.html'),
+        name='monitor-api-stats'),
     url(r'^totals_stats_ajax/$', monitor.views.totals_stats_ajax,
         name='monitor-totals-stats-ajax'),
     url(r'^ajax_tags_stats/$', monitor.views.tags_stats_ajax,
         name='monitor-tags-stats-ajax'),
     url(r'^ajax_queries_stats/$', monitor.views.queries_stats_ajax,
         name='monitor-queries-stats-ajax'),
+    url(r'^ajax_api_usage_stats/$', monitor.views.api_usage_stats_ajax,
+        name='monitor-api-stats-ajax'),
     url(r'^ajax_downloads_stats/$', monitor.views.downloads_stats_ajax,
         name='monitor-downloads-stats-ajax'),
     url(r'^ajax_donations_stats/$', monitor.views.donations_stats_ajax,
