+++ templates/monitor/api_usage.html
@@ -1,0 +1,72 @@
+{% extends "base.html" %}
+
+{% block head %}
+{{ block.super }}
+<style type="text/css">
+  .axis path,.axis line {
+    fill: none;
+    stroke: #b6b6b6;
+    shape-rendering: crispEdges;
+  }
+  g.legend { 
+    font: 10px sans-serif;
+  }
+  .legend-item{  
+    cursor: pointer;
+  }
+  .api-usage{
+    min-height: 200px;
+  }
+</style>
+{% endblock head %}
+
+{% block title %}
+    Monitor
+{% endblock title %}
+
+{% block content %}
+
+    <h2>Api usage</h2>
+    <h3>Api usage by day on last two weeks</h3>
+    <div class='api-usage'>
+    </div>
+
+    <script>
+      var queriesApiUrl = '{% url 'monitor-api-stats-ajax' %}';
+      $.get(queriesApiUrl, function(d){
+        var usage = [];
+        for (var key in d.results) {
+          if (d.results.hasOwnProperty(key)) {
+            usage.push({day: parseInt(key)*1000, count: d.results[key]});
+          }
+        }
+        if (usage.length > 0) {
+          displayCharts('.api-usage', [usage], {
+              yText: 'Request',
+              attrX: 'day',
+              attrY: 'count',
+              timeFormat: "%a %d",
+              tickEvery: d3.timeDay.every(1),
+              legendData: [{color: 'crimson', name: 'requests'}]
+            }, {});
+        } else {
+          $('.api-usage').html('<p>There are no requests in the given period.</p>')
+        }
+      });
+
+    </script>
+    <script src="https://d3js.org/d3.v4.min.js" type="text/javascript"></script>
+    <script src="https://d3js.org/d3-axis.v1.min.js"></script>
+    <script src="https://d3js.org/d3-array.v1.min.js"></script>
+    <script src="https://d3js.org/d3-collection.v1.min.js"></script>
+    <script src="https://d3js.org/d3-color.v1.min.js"></script>
+    <script src="https://d3js.org/d3-format.v1.min.js"></script>
+    <script src="https://d3js.org/d3-interpolate.v1.min.js"></script>
+    <script src="https://d3js.org/d3-time.v1.min.js"></script>
+    <script src="https://d3js.org/d3-time-format.v2.min.js"></script>
+    <script src="https://d3js.org/d3-scale.v1.min.js"></script>
+    <script src="https://d3js.org/d3-path.v1.min.js"></script>
+    <script src="https://d3js.org/d3-shape.v1.min.js"></script>
+    <script src="{{media_url}}js/wordcloud2.js"></script>
+    <script src="{{media_url}}js/chart.js" type="text/javascript"></script>
+{% endblock content %}
