+++ templates/monitor/stats.html
@@ -148,5 +148,6 @@
     <script src="https://d3js.org/d3-path.v1.min.js"></script>
     <script src="https://d3js.org/d3-shape.v1.min.js"></script>
     <script src="{{media_url}}js/wordcloud2.js"></script>
+    <script src="{{media_url}}js/chart.js" type="text/javascript"></script>
     <script src="{{media_url}}js/stats.js" type="text/javascript"></script>
 {% endblock content %}
